refactor(SCRUM-235) : 디렉토리 정리
