+++ src/main/java/com/team_nebula/nebula/domain/star/search/batch/ElasticsearchBatchService.java
@@ -1,4 +1,4 @@
-package com.team_nebula.nebula.domain.star.search.service;
+package com.team_nebula.nebula.domain.star.search.batch;
 
 import co.elastic.clients.elasticsearch.ElasticsearchClient;
 import co.elastic.clients.elasticsearch.core.BulkRequest;
@@ -8,9 +8,11 @@ import com.google.common.collect.Lists;
 import com.team_nebula.nebula.domain.star.search.document.StarSearchDocument;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.concurrent.CompletableFuture;
 
 @Service
 @RequiredArgsConstructor
@@ -18,6 +20,17 @@ import java.util.List;
 public class ElasticsearchBatchService {
 
     private final ElasticsearchClient elasticsearchClient;
+
+    @Async("batchExecutor")
+    public CompletableFuture<Void> processBatchAsync(List<StarSearchDocument> documents) {
+        try {
+            batchIndexDocuments(documents);
+            return CompletableFuture.completedFuture(null);
+        } catch (Exception e) {
+            log.error("Async batch processing failed", e);
+            return CompletableFuture.failedFuture(e);
+        }
+    }
 
     public void batchIndexDocuments(List<StarSearchDocument> documents) {
         if (documents.isEmpty()) return;
