+++ src/main/java/com/team_nebula/nebula/domain/star/search/cache/CacheInvalidationService.java
@@ -1,0 +1,4 @@
+package com.team_nebula.nebula.domain.star.search.cache;
+
+public class CacheInvalidationService {
+}
